Create a copy of the logger when being changed
Just like logrus itself we don't modify the existing logger when calling
WithName, WithValues or V, instead we create a copy and modify that.
This will ensure that a logger passed to a sub package will still be
usable in the main package if modified.

This also solves a race condition when modifying the name (Closes #2)