Stop overwriting logger on struct when adding fields
When adding caller, and for Info logs any other field, don't overwrite the
logger on the logrusr struct directly since we only want each individual
message to contain this information.

To make this easier, also always treat the logger internally as a
`logrus.Entry` so we don't need to type cast when using it. This also
makes us able to use the public duplication interface when duplicating
the logger.